+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.1
+### **Enhancements**
+* Default location for CSV input files is now used - `input/records.csv`
+* For DataDog metrics, publish a 0 for what did NOT happen (failure, success and error) - this makes math easier in the reports (as otherwise we have N/As, which cannot be using formulas)
+
 ## 0.4.0
 ### **Breaking changes**
 * Tokens are now kept in the config file - property: `virtualDevices`. See [README](README.md#virtual-device-setup).
@@ -5,7 +10,7 @@
 * All output files are now kept in the /output directory - such as /output/results.csv.
 * Reporting uses new `customer` property for the configuration file - this is a REQUIRED field
 
-### **Other changes**
+### **Enhancements**
 * Ability to tag tokens to restrict processing of records to certain tokens - [read more](README.md#virtual-device-setup)
 
 ## 0.3.0